+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Removed
+- Following pacemaker 3, support for nagios and upstart resource classes has
+  been removed ([RHEL-49520])
+
 ### Added
 - Support for output formats `json` and `cmd` to `pcs tag config` command
   ([RHEL-21047])
@@ -43,6 +47,7 @@
 [RHEL-21047]: https://issues.redhat.com/browse/RHEL-21047
 [RHEL-33386]: https://issues.redhat.com/browse/RHEL-33386
 [RHEL-44432]: https://issues.redhat.com/browse/RHEL-44432
+[RHEL-49520]: https://issues.redhat.com/browse/RHEL-49520
 [RHEL-55723]: https://issues.redhat.com/browse/RHEL-55723
 [RHEL-61747]: https://issues.redhat.com/browse/RHEL-61747
 [RHEL-61889]: https://issues.redhat.com/browse/RHEL-61889